FEATURE: Create `CoreDataStack` to handle `persistentContainer` initialization; Clean code
